pyproject.toml
[project]
name = "avital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "python-bidi>=0.6.7",
]
